Fixed decimal/double issue

+++ API/UserCustomIndices/Model/Response/EarningsGrowth.cs
@@ -3,7 +3,7 @@
     public struct EarningsGrowth
     {
         public int TimePeriod;
-        public decimal Upper;
-        public decimal Lower;
+        public double Upper;
+        public double Lower;
     }
 }
